Refactoring
Also fixed little bug with datespan which is not include last entry
Made more files with packaging classes and functions more semantically
Added typehints
Added test which compare entries of src and dst
