+++ etl/utils.py
@@ -1,0 +1,7 @@
+from datetime import date, datetime, timedelta
+
+def datespan(startDate, endDate, delta=timedelta(hours=1)):
+    currentDate = startDate
+    while currentDate < endDate:
+        yield currentDate
+        currentDate += delta
