+++ tests/test_etl_contest.py
@@ -1,6 +1,7 @@
 import pymysql
 from .helpers import ping_container
 
+import etl
 
 def test_container_is_alive(mysql_source_image):
     assert ping_container(mysql_source_image)
@@ -45,12 +46,44 @@ def test_containers_assets_is_ready(mysql_source_image,
 def test_data_transfer(mysql_source_image,
                        mysql_destination_image):
     """
-
     :param mysql_source_image: Контейнер mysql-источника с исходными данными
     :param mysql_destination_image: Контейнер mysql-назначения
     :return:
     """
 
-    #   put your code for testing here!
 
-    pass
+    CHECK_DST = """
+        SELECT
+            t.id, t.dt, t.idoper, t.move, t.amount, t.name_oper
+        FROM
+            transactions_denormalized t
+        ORDER BY t.dt
+    """
+
+    CHECK_SRC = """
+        SELECT
+            t.id, t.dt, t.idoper, t.move, t.amount, o.name as name_oper
+        FROM
+            transactions t
+        JOIN
+            operation_types o
+        ON
+            t.idoper = o.id
+        ORDER BY t.dt
+    """
+
+    etl.process(mysql_source_image, mysql_destination_image)
+
+    with pymysql.connect(**mysql_destination_image) as cdst:
+                with cdst.cursor() as c:
+                    c.execute(CHECK_DST)
+                    dst_table = list(c.fetchall())
+
+    with pymysql.connect(**mysql_source_image) as csrc:
+                with csrc.cursor() as c:
+                    c.execute(CHECK_SRC)
+                    src_table = list(c.fetchall())
+
+    import pprint
+
+    assert dst_table == src_table
